requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.31
psycopg2-binary==2.9.9
alembic==1.13.2
python-jose[cryptography]==3.3.0
bcrypt==4.1.3
redis==5.0.7
pydantic==2.7.4
pydantic-settings==2.3.4
python-multipart==0.0.9
httpx==0.27.0
pytest==8.2.2
pytest-asyncio==0.23.7
pytest-cov==5.0.0
fakeredis==2.34.1
anyio==4.4.0
email-validator==2.2.0
python-dotenv==1.0.1
structlog==24.2.0
